Draw airplane on canvas
Creates new scripts.js file in new javascripts dir in lib. A suggestion
for organizing js and css code for this project.

Current plane is only in red. Sizing and new colors can be created by
changing values of parameters. These should be turned into objects so
that they can be created at will once we are happy with the presentation.